MedicalMigration/scripts/requirements.txt
pandas==2.0.3
numpy==1.26.0
pymongo==4.6.1
python-dotenv==1.0.1

